currently CoreDeployTestFixture usage creates testing db leak with db handles
